Create and link CSS file, add internal stylings
Create an external stylesheet as styles.css and link it with the
index.html file in the 'head' section using <link> tag. Add internal
style to the paragraph using <style> tag under the 'head' section.
Add inline style to the button using 'style' attribute in the opening
tag of <button>